requirements.txt
RPi.GPIO==0.7.0
gpiozero==1.6.2
pygame==2.1.2